Rename License to VLicense
