+++ src/components/VAudioTrack/layouts/VRowLayout.vue
@@ -88,7 +88,7 @@ import { computed, defineComponent, useContext } from '@nuxtjs/composition-api'
 import { useBrowserIsBlink } from '~/composables/use-browser-detection'
 
 import VAudioThumbnail from '~/components/VAudioThumbnail/VAudioThumbnail.vue'
-import VLicense from '~/components/License/VLicense.vue'
+import VLicense from '~/components/VLicense/VLicense.vue'
 import VLink from '~/components/VLink.vue'
 
 export default defineComponent({
