+++ src/components/VLicense/meta/VLicense.stories.mdx
@@ -6,7 +6,7 @@ import {
   Story,
 } from '@storybook/addon-docs'
 
-import VLicense from '~/components/License/VLicense.vue'
+import VLicense from '~/components/VLicense/VLicense.vue'
 
 <Meta title="Components/VLicense" component={VLicense} />
